+++ ryu/BUILD
@@ -3,13 +3,16 @@ package(default_visibility=["//visibility:public"])
 cc_library(
   name = "ryu",
   srcs = [
+    "f2parts.c",
     "f2s.c",
+    "d2parts.c",
     "d2s.c",
     "d2s.h",
     "d2s_full_table.h",
     "digit_table.h",
     "mulshift128.h",
     "common.h",
+    "to_decimal.h",
   ],
   hdrs = ["ryu.h"],
 )
