+++ ryu/f2parts.c
@@ -1,0 +1,299 @@
+// Copyright 2018 Ulf Adams and contributors
+//
+// The contents of this file may be used under the terms of the Apache License,
+// Version 2.0.
+//
+//    (See accompanying file LICENSE-Apache or copy at
+//     http://www.apache.org/licenses/LICENSE-2.0)
+//
+// Alternatively, the contents of this file may be used under the terms of
+// the Boost Software License, Version 1.0.
+//    (See accompanying file LICENSE-Boost or copy at
+//     https://www.boost.org/LICENSE_1_0.txt)
+//
+// Unless required by applicable law or agreed to in writing, this software
+// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.
+
+// Runtime compiler options:
+// -DRYU_DEBUG Generate verbose debugging output to stdout.
+
+#include "ryu/ryu.h"
+
+#include <assert.h>
+#include <stdbool.h>
+#include <stdint.h>
+#include <stdlib.h>
+#include <string.h>
+#include <limits.h>
+
+#include "ryu/common.h"
+
+#ifndef NO_DIGIT_TABLE
+#include "ryu/digit_table.h"
+#endif
+
+#ifdef RYU_DEBUG
+#include <stdio.h>
+#endif
+
+#define FLOAT_MANTISSA_BITS 23
+#define FLOAT_EXPONENT_BITS 8
+
+// This table is generated by PrintFloatLookupTable.
+#define FLOAT_POW5_INV_BITCOUNT 59
+static const uint64_t FLOAT_POW5_INV_SPLIT[31] = {
+  576460752303423489u, 461168601842738791u, 368934881474191033u, 295147905179352826u,
+  472236648286964522u, 377789318629571618u, 302231454903657294u, 483570327845851670u,
+  386856262276681336u, 309485009821345069u, 495176015714152110u, 396140812571321688u,
+  316912650057057351u, 507060240091291761u, 405648192073033409u, 324518553658426727u,
+  519229685853482763u, 415383748682786211u, 332306998946228969u, 531691198313966350u,
+  425352958651173080u, 340282366920938464u, 544451787073501542u, 435561429658801234u,
+  348449143727040987u, 557518629963265579u, 446014903970612463u, 356811923176489971u,
+  570899077082383953u, 456719261665907162u, 365375409332725730u
+};
+#define FLOAT_POW5_BITCOUNT 61
+static const uint64_t FLOAT_POW5_SPLIT[47] = {
+ 1152921504606846976u, 1441151880758558720u, 1801439850948198400u, 2251799813685248000u,
+ 1407374883553280000u, 1759218604441600000u, 2199023255552000000u, 1374389534720000000u,
+ 1717986918400000000u, 2147483648000000000u, 1342177280000000000u, 1677721600000000000u,
+ 2097152000000000000u, 1310720000000000000u, 1638400000000000000u, 2048000000000000000u,
+ 1280000000000000000u, 1600000000000000000u, 2000000000000000000u, 1250000000000000000u,
+ 1562500000000000000u, 1953125000000000000u, 1220703125000000000u, 1525878906250000000u,
+ 1907348632812500000u, 1192092895507812500u, 1490116119384765625u, 1862645149230957031u,
+ 1164153218269348144u, 1455191522836685180u, 1818989403545856475u, 2273736754432320594u,
+ 1421085471520200371u, 1776356839400250464u, 2220446049250313080u, 1387778780781445675u,
+ 1734723475976807094u, 2168404344971008868u, 1355252715606880542u, 1694065894508600678u,
+ 2117582368135750847u, 1323488980084844279u, 1654361225106055349u, 2067951531382569187u,
+ 1292469707114105741u, 1615587133892632177u, 2019483917365790221u
+};
+
+static inline int32_t pow5Factor(uint32_t value) {
+  for (int32_t count = 0; value > 0; ++count) {
+    if (value % 5 != 0) {
+      return count;
+    }
+    value /= 5;
+  }
+  return 0;
+}
+
+// Returns true if value is divisible by 5^p.
+static inline bool multipleOfPowerOf5(const uint32_t value, const int32_t p) {
+  return pow5Factor(value) >= p;
+}
+
+// It seems to be slightly faster to avoid uint128_t here, although the
+// generated code for uint128_t looks slightly nicer.
+static inline uint32_t mulShift(const uint32_t m, const uint64_t factor, const int32_t shift) {
+  assert(shift > 32);
+
+  // The casts here help MSVC to avoid calls to the __allmul library
+  // function.
+  const uint32_t factorLo = (uint32_t)(factor);
+  const uint32_t factorHi = (uint32_t)(factor >> 32);
+  const uint64_t bits0 = (uint64_t)m * factorLo;
+  const uint64_t bits1 = (uint64_t)m * factorHi;
+
+#if defined(_M_IX86) || defined(_M_ARM)
+  // On 32-bit platforms we can avoid a 64-bit shift-right since we only
+  // need the upper 32 bits of the result and the shift value is > 32.
+  const uint32_t bits0Hi = (uint32_t)(bits0 >> 32);
+  uint32_t bits1Lo = (uint32_t)(bits1);
+  uint32_t bits1Hi = (uint32_t)(bits1 >> 32);
+  bits1Lo += bits0Hi;
+  bits1Hi += (bits1Lo < bits0Hi);
+  const int32_t s = shift - 32;
+  return (bits1Hi << (32 - s)) | (bits1Lo >> s);
+#else
+  const uint64_t sum = (bits0 >> 32) + bits1;
+  const uint64_t shiftedSum = sum >> (shift - 32);
+  assert(shiftedSum <= UINT32_MAX);
+  return (uint32_t) shiftedSum;
+#endif
+}
+
+static inline uint32_t mulPow5InvDivPow2(const uint32_t m, const uint32_t q, const int32_t j) {
+  return mulShift(m, FLOAT_POW5_INV_SPLIT[q], j);
+}
+
+static inline uint32_t mulPow5divPow2(const uint32_t m, const uint32_t i, const int32_t j) {
+  return mulShift(m, FLOAT_POW5_SPLIT[i], j);
+}
+
+struct fparts f2parts(float f) {
+  // Step 1: Decode the floating-point number, and unify normalized and subnormal cases.
+  const uint32_t mantissaBits = FLOAT_MANTISSA_BITS;
+  const uint32_t exponentBits = FLOAT_EXPONENT_BITS;
+  const uint32_t offset = (1u << (exponentBits - 1)) - 1;
+
+  uint32_t bits = 0;
+  // This only works on little-endian architectures.
+  memcpy(&bits, &f, sizeof(float));
+
+  // Decode bits into sign, mantissa, and exponent.
+  const bool sign = ((bits >> (mantissaBits + exponentBits)) & 1) != 0;
+  const uint32_t ieeeMantissa = bits & ((1u << mantissaBits) - 1);
+  const uint32_t ieeeExponent = (uint32_t) ((bits >> mantissaBits) & ((1u << exponentBits) - 1));
+
+#ifdef RYU_DEBUG
+  printf("IN=");
+  for (int32_t bit = 31; bit >= 0; --bit) {
+    printf("%d", (bits >> bit) & 1);
+  }
+  printf("\n");
+#endif
+
+  int32_t e2;
+  uint32_t m2;
+  // Case distinction; exit early for the easy cases.
+  if (ieeeExponent == ((1u << exponentBits) - 1u) || (ieeeExponent == 0 && ieeeMantissa == 0)) {
+    return (struct fparts) {
+      .sign = sign,
+      .exp = ieeeExponent ? INT16_MAX : 0,
+      .output = ieeeMantissa,
+    };
+  } else if (ieeeExponent == 0) {
+    // We subtract 2 so that the bounds computation has 2 additional bits.
+    e2 = 1 - offset - mantissaBits - 2;
+    m2 = ieeeMantissa;
+  } else {
+    e2 = ieeeExponent - offset - mantissaBits - 2;
+    m2 = (1u << mantissaBits) | ieeeMantissa;
+  }
+  const bool even = (m2 & 1) == 0;
+  const bool acceptBounds = even;
+
+#ifdef RYU_DEBUG
+  printf("S=%s E=%d M=%u\n", sign ? "-" : "+", e2, m2);
+#endif
+
+  // Step 2: Determine the interval of legal decimal representations.
+  const uint32_t mv = 4 * m2;
+  const uint32_t mp = 4 * m2 + 2;
+  const uint32_t mm = 4 * m2 - (((m2 != (1u << mantissaBits)) || (ieeeExponent <= 1)) ? 2 : 1);
+
+  // Step 3: Convert to a decimal power base using 64-bit arithmetic.
+  uint32_t vr;
+  uint32_t vp, vm;
+  int32_t e10;
+  bool vmIsTrailingZeros = false;
+  bool vrIsTrailingZeros = false;
+  uint8_t lastRemovedDigit = 0;
+  if (e2 >= 0) {
+    const int32_t q = log10Pow2(e2);
+    e10 = q;
+    const int32_t k = FLOAT_POW5_INV_BITCOUNT + pow5bits(q) - 1;
+    const int32_t i = -e2 + q + k;
+    vr = mulPow5InvDivPow2(mv, q, i);
+    vp = mulPow5InvDivPow2(mp, q, i);
+    vm = mulPow5InvDivPow2(mm, q, i);
+#ifdef RYU_DEBUG
+    printf("%d * 2^%d / 10^%d\n", mv, e2, q);
+    printf("V+=%d\nV =%d\nV-=%d\n", vp, vr, vm);
+#endif
+    if (q != 0 && ((vp - 1) / 10 <= vm / 10)) {
+      // We need to know one removed digit even if we are not going to loop below. We could use
+      // q = X - 1 above, except that would require 33 bits for the result, and we've found that
+      // 32-bit arithmetic is faster even on 64-bit machines.
+      const int32_t l = FLOAT_POW5_INV_BITCOUNT + pow5bits(q - 1) - 1;
+      lastRemovedDigit = (uint8_t) (mulPow5InvDivPow2(mv, q - 1, -e2 + q - 1 + l) % 10);
+    }
+    if (q <= 9) {
+      // Only one of mp, mv, and mm can be a multiple of 5, if any.
+      if (mv % 5 == 0) {
+        vrIsTrailingZeros = multipleOfPowerOf5(mv, q);
+      } else {
+        if (acceptBounds) {
+          vmIsTrailingZeros = multipleOfPowerOf5(mm, q);
+        } else {
+          vp -= multipleOfPowerOf5(mp, q);
+        }
+      }
+    }
+  } else {
+    const int32_t q = log10Pow5(-e2);
+    e10 = q + e2;
+    const int32_t i = -e2 - q;
+    const int32_t k = pow5bits(i) - FLOAT_POW5_BITCOUNT;
+    int32_t j = q - k;
+    vr = mulPow5divPow2(mv, i, j);
+    vp = mulPow5divPow2(mp, i, j);
+    vm = mulPow5divPow2(mm, i, j);
+#ifdef RYU_DEBUG
+    printf("%d * 5^%d / 10^%d\n", mv, -e2, q);
+    printf("%d %d %d %d\n", q, i, k, j);
+    printf("V+=%d\nV =%d\nV-=%d\n", vp, vr, vm);
+#endif
+    if (q != 0 && ((vp - 1) / 10 <= vm / 10)) {
+      j = q - 1 - (pow5bits(i + 1) - FLOAT_POW5_BITCOUNT);
+      lastRemovedDigit = (uint8_t) (mulPow5divPow2(mv, i + 1, j) % 10);
+    }
+    if (q <= 1) {
+      vrIsTrailingZeros = (~mv & 1) >= (uint32_t) q;
+      if (acceptBounds) {
+        vmIsTrailingZeros = (~mm & 1) >= (uint32_t) q;
+      } else {
+        --vp;
+      }
+    } else if (q < 31) { // TODO(ulfjack): Use a tighter bound here.
+      vrIsTrailingZeros = (mv & ((1u << (q - 1)) - 1)) == 0;
+    }
+  }
+#ifdef RYU_DEBUG
+  printf("e10=%d\n", e10);
+  printf("V+=%u\nV =%u\nV-=%u\n", vp, vr, vm);
+  printf("vm is trailing zeros=%s\n", vmIsTrailingZeros ? "true" : "false");
+  printf("vr is trailing zeros=%s\n", vrIsTrailingZeros ? "true" : "false");
+#endif
+
+  // Step 4: Find the shortest decimal representation in the interval of legal representations.
+  uint32_t removed = 0;
+  uint32_t output;
+  if (vmIsTrailingZeros || vrIsTrailingZeros) {
+    // General case, which happens rarely.
+    while (vp / 10 > vm / 10) {
+      vmIsTrailingZeros &= vm % 10 == 0;
+      vrIsTrailingZeros &= lastRemovedDigit == 0;
+      lastRemovedDigit = (uint8_t) (vr % 10);
+      vr /= 10;
+      vp /= 10;
+      vm /= 10;
+      ++removed;
+    }
+    if (vmIsTrailingZeros) {
+      while (vm % 10 == 0) {
+        vrIsTrailingZeros &= lastRemovedDigit == 0;
+        lastRemovedDigit = (uint8_t) (vr % 10);
+        vr /= 10;
+        vp /= 10;
+        vm /= 10;
+        ++removed;
+      }
+    }
+    if (vrIsTrailingZeros && (lastRemovedDigit == 5) && (vr % 2 == 0)) {
+      // Round down not up if the number ends in X50000.
+      lastRemovedDigit = 4;
+    }
+    // We need to take vr+1 if vr is outside bounds or we need to round up.
+    output = vr +
+        ((vr == vm && (!acceptBounds || !vmIsTrailingZeros)) || (lastRemovedDigit >= 5));
+  } else {
+    // Common case.
+    while (vp / 10 > vm / 10) {
+      lastRemovedDigit = (uint8_t) (vr % 10);
+      vr /= 10;
+      vp /= 10;
+      vm /= 10;
+      ++removed;
+    }
+    // We need to take vr+1 if vr is outside bounds or we need to round up.
+    output = vr + ((vr == vm) || (lastRemovedDigit >= 5));
+  }
+  struct fparts parts;
+  parts.sign = sign;
+  parts.output = output;
+  parts.exp = e10 + removed;
+
+  return parts;
+}
